Refactor code, Reorganize folder closes #37 closes #36 closes #24
