+++ .gitignore
@@ -12,7 +12,7 @@
 /build
 
 #API KEYS
-*/API_KEY
+movie_api_key.js
 
 # misc 
 .DS_Store
