+++ src/components/Home/Home.js
@@ -1,17 +1,12 @@
 import React from 'react'
-import Nav from '../../containers/Nav/Nav'
-import View from '../../containers/View/View'
+import { Nav, View } from '../'
 import './Home.scss'
 
-const Home = () => {
-  // ternary check to see if user is logged in 
-  // if so, return home, if not, Redirect to login
-  return (
-    <div className='Home' >
-      <Nav />
-      <View />
-    </div >
-  )
-}
+const Home = () => (
+  <div className='Home' >
+    <Nav />
+    <View />
+  </div >
+)
 
 export default Home;
