+++ src/components/Movie/Movie.scss
