+++ src/components/Popup/Popup.js
@@ -1,12 +1,10 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-
 const Popup = ({ movie, user, setLoginPrompt }) => {
   const imgUrl = 'https://image.tmdb.org/t/p/w500'
 
   return (
-
     <div>
       <Link to='/'>x</Link>
       <h1 className='movie-title'>{movie.title}</h1>
@@ -19,7 +17,6 @@ const Popup = ({ movie, user, setLoginPrompt }) => {
       <button className='favorite-btn' onClick={() => setLoginPrompt(!user)}>Add To Favorites </button>
 
     </div>)
-
 }
 
 export default Popup
