+++ src/components/index.js
@@ -1,0 +1,13 @@
+import Home from './Home/Home'
+import Movie from './Movie/Movie'
+import Nav from './Nav/Nav'
+import Popup from './Popup/Popup'
+import View from './View/View'
+
+export {
+  Home,
+  Movie,
+  Nav,
+  Popup,
+  View
+}
