+++ src/containers/App/App.spec.js
