--- src/containers/Button/Button.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const Button = (prop) => {
-  return 
-}
