+++ src/containers/Carousel/Carousel.scss
