+++ src/containers/Carousel/Carousel.spec.js
