+++ src/containers/Profile/Profile.js
@@ -1,26 +1,29 @@
 import React from 'react'
+import { Link } from 'react-router-dom'
+import { connect } from 'react-redux'
+import { setUser } from '../../actions'
 import './Profile.scss'
-import {Link} from 'react-router-dom'
-import {connect} from 'react-redux'
-import {setUser} from '../../actions'
 
-const Profile = (props)=>{
-  let shouldRedirect = false;
-  const handleClick=()=>{
-    shouldRedirect = true;
-  }
+const Profile = (props) => {
 
   return (
-      <div className='Profile' >
-        Profile
-        <Link onClick={handleClick} to='/login'>Login</Link>
-
-        <button onClick={()=>props.setUser(null)}>Logout</button>
-      </div>)
+    <div className='Profile' >
+      <p>Profile</p>
+      {!props.user ?
+        <Link to='/login'>Login</Link>
+        :
+        <button onClick={() => props.setUser(null)}>Logout</button>}
+    </div>
+  )
 }
 
-export const mapDispatchToProps = (dispatch)=>({
-  setUser:(user)=>dispatch(setUser(user))
+export const mapStateToProps = (state) => ({
+  user: state.user
 })
 
-export default connect(null,mapDispatchToProps)(Profile);
+export const mapDispatchToProps = (dispatch) => ({
+  setUser: (user) => dispatch(setUser(user))
+})
+
+
+export default connect(mapStateToProps, mapDispatchToProps)(Profile);
