+++ src/containers/index.js
@@ -1,0 +1,14 @@
+import App from './App/App'
+import Carousel from './Carousel/Carousel'
+import CreateUser from './CreateUser/CreateUser'
+import Login from './Login/Login'
+import Profile from './Profile/Profile'
+
+
+export {
+  Profile,
+  Carousel,
+  CreateUser,
+  Login,
+  App
+}
