+++ src/styles/main.scss
@@ -1,3 +1,3 @@
 @import '/normalize';
 @import './variables';
-@import './components/App/App.scss';
+@import '.././containers/App/App.scss';
